[ADD] sale_order_extended: view file created for displaying tree view of sale.order.line
order_lines_views.xml file created to display the tree view of records of sale.order.line, only those records which are not in
cancel or done state.

[MOD] sale_order_extended: added button in sale order and created method in sale.order and added xml file in manifest

In sale_order.py file added a method to filter the sale.order.line records which are in the opened sale order and whose state should not be
done or cancel. In sale_order.xml file added a button-box to display the sale.order.line records.