[MOD] sale_order_extended : inherited filter of sale order and added a filter and defined a search method in sale order
Inherited filter of sale order and added a method in sale order for filter only that records of sale order of which pickings are either
in done state or cancel state.

[FIX] sale_order_extended: when product template is changed the varients of that template shoud be auto added in sale order line

There is a many2many field named product template in sale order, when the template is changed then its variants should be automatically
be added in sale order line. Was getting singleton error in product which is fixed now.